Increase bit width of priority index to 64

+++ lib/data_structures.h
@@ -28,7 +28,7 @@ typedef struct ProcessData
     int pid;
     int t_arrival;
     int t_running;
-    int priority;
+    long long priority;
 } ProcessData;
 
 /**
